Use raw github instead monkeytype server
