+++ README.md
@@ -29,10 +29,11 @@ typer --min-word-length 5
 There is no maximum value, but anything below 1 will count as no min length.
 
 You can use Monkeytype as a source of words, just pass `-m, --monkeytype` flag,
-by default it'll use `english` dictionary, you can change that by adding `--monkeytype-language string` additionally.
+by default it'll use `english` dictionary, you can change that by adding `--monkeytype-language string` flag.
 ```
-typer -m --monkeytype-language english_1k
+typer -m --monkeytype-language code_go
 ```
+see: [monkeytype/languages/_list.json](https://github.com/monkeytypegame/monkeytype/blob/master/frontend/static/languages/_list.json)
 
 If you want to provide your own text, you can pass in a file name with the `--file` flag. The typing test will use the contents of the specified file.
 ```
@@ -56,6 +57,7 @@ theme:
     #gradient: '#ff0000' # if passed, will generate a gradient from previous color to this one
   graph:
     # see: https://pkg.go.dev/github.com/guptarohit/asciigraph#AnsiColor
+    # see: https://github.com/guptarohit/asciigraph/blob/master/color.go#L152-L292
     color: blue # does not use rgb but rather ANSI codes
     height: 3   # height of the graph
   text:
