+++ pkg/theme/theme.go
@@ -4,7 +4,6 @@ import (
 	"github.com/charmbracelet/bubbles/progress"
 	"github.com/guptarohit/asciigraph"
 	"github.com/muesli/termenv"
-	"github.com/spf13/viper"
 )
 
 func DefaultTheme() *Theme {
@@ -32,22 +31,6 @@ func DefaultTheme() *Theme {
 			Height: 3,
 		},
 	}
-}
-
-func LoadViper(v *viper.Viper, first bool) (*Theme, error) {
-	theme := DefaultTheme()
-	v.UnmarshalKey("theme", theme)
-	if !first || theme.File == "" {
-		return theme, nil
-	}
-
-	v = viper.New()
-	v.SetConfigFile(theme.File)
-	err := v.ReadInConfig()
-	if err != nil {
-		return theme, err
-	}
-	return LoadViper(v, false)
 }
 
 type Theme struct {
